Syncing down changed from other forks

+++ build_jzmq.sh
@@ -6,9 +6,13 @@ url='https://github.com/nathanmarz/jzmq.git'
 buildroot=build
 fakeroot=jzmq
 origdir=$(pwd)
+prefix="/usr"
 description='JZMQ is the Java bindings for ZeroMQ'
 # Make sure JAVA_HOME is set. Uncomment if necessary
-export JAVA_HOME='/usr/lib/jvm/java-6-sun-1.6.0.31'
+if [ "${JAVA_HOME}x" == "x" ]; then
+  echo Please set JAVA_HOME before running.
+  exit -1
+fi
 
 #_ MAIN _#
 rm -rf ${name}*.deb
@@ -17,8 +21,15 @@ rm -rf ${buildroot}
 mkdir -p ${buildroot}
 git clone ${url}
 cd jzmq
+sed -i 's/classdist_noinst.stamp/classnoinst.stamp/g' src/Makefile.am
 ./autogen.sh
-./configure
+./configure --prefix=${prefix}
+# https://github.com/zeromq/jzmq/issues/114#issuecomment-6927797
+touch src/classdist_noinst.stamp
+cd src/
+CLASSPATH=.:./.:$CLASSPATH
+javac -d . org/zeromq/ZMQ.java org/zeromq/App.java org/zeromq/ZMQForwarder.java org/zeromq/EmbeddedLibraryTools.java org/zeromq/ZMQQueue.java org/zeromq/ZMQStreamer.java org/zeromq/ZMQException.java
+cd ..
 make
 make install DESTDIR=${origdir}/${buildroot}
